shopinvader_locomotive: allow hybrid setup
In a hybrid setup, you may have locomotive and non locomotive
backends. This commit introduce an "is_locomotive" flag on
the backend.

+++ shopinvader_locomotive/component/event_listeners.py
@@ -14,6 +14,8 @@ class ShopinvaderBindingListener(Component):
 
     @skip_if(lambda self, record, **kwargs: self.no_connector_export(record))
     def on_record_create(self, record, fields=None):
+        if not record.backend_id.is_locomotive:
+            return
         record.with_delay().export_record(_fields=fields)
 
     def _get_export_not_triggered_fields(self):
@@ -21,6 +23,8 @@ class ShopinvaderBindingListener(Component):
 
     @skip_if(lambda self, record, **kwargs: self.no_connector_export(record))
     def on_record_write(self, record, fields=None):
+        if not record.backend_id.is_locomotive:
+            return
         if not fields:
             return
         # skip export if updating only fields that are not relevant
@@ -30,6 +34,8 @@ class ShopinvaderBindingListener(Component):
         record.with_delay().export_record(_fields=fields)
 
     def on_record_unlink(self, record):
+        if not record.backend_id.is_locomotive:
+            return
         with record.backend_id.work_on(record._name) as work:
             external_id = work.component(usage="binder").to_external(record)
             if external_id:
@@ -61,4 +67,6 @@ class ShopinvaderRecordListener(Component):
         if "shopinvader_bind_ids" not in record._fields:
             return
         for binding in record.shopinvader_bind_ids:
+            if not binding.backend_id.is_locomotive:
+                continue
             binding.unlink()
